Refactor current page link logic & raise it to a reusable module
